Fix UnitBuff deprecation impact on memory usage (garbage leak) when the check for stealth option is used.
In 10.2.5 UnitBuff and acolytes were deprecated and are falling back to calling the new C_UnitAuras API functions which create a new Lua table each time they return something and thus, leak memory... (if only we could provide those functions with a table to use... *sigh*)

This is a work in progress, more work is needed. In 10.2.5, Decursive will still leak memory as soon as there are debuffs on units... This update fix the idle case which was probably the worst since buffs are long lasting and on all units all the time.