work on issue #20: Add settings in MicroUnit Frames performance options to control the global periodic debuff rescan
We need to test if we can go without this periodic rescan, you can set the setting to 0 in order to disable it completely or increase the delay between scans to a high value (10s) and enable the "Periodic scan debug reporting" options. This will create a debug report that would pop out after a fight if the scan function detected something that was not detected by the event system.
NOTE: Currently changing the curing options (directly or indirectly through ability changes may trigger this report - next alpha versions will prevent this)