Fix for broken talent in 1.15.8.63829 (WoW Classic)
Remove talent availability check entirely as Blizzard broke talents in a Friday night patch and won't fix it until Monday. This talent availability check is probably no longer necessary anyway. We still wait for the player's GUID to be available though, just in case.

This talent issue also triggered a nil error when attempting to display the priority list due to uninitialized class color table, use a default color in that case.

#skipretail #skipmop